requirements.txt
Django==3.2.10
PyNaCl==1.4.0
base58==2.1.1